refactor:separate block registration from editor component
