+++ src/custom-blocks/allowed-third-party-embed/index.js
@@ -1,28 +1,21 @@
 /**
  * Allowed third party embed
  */
+
 import { __ } from "@wordpress/i18n";
 import { registerBlockType } from "@wordpress/blocks";
-import {
-  RichText,
-  InspectorControls,
-  BlockControls,
-} from "@wordpress/block-editor";
-import {
-  SandBox,
-  TextareaControl,
-  ToolbarButton,
-  ToolbarGroup,
-} from "@wordpress/components";
-import { useState } from "@wordpress/element";
+
+import Edit from "./edit";
+
 registerBlockType("wb-blocks/allowed-third-party-embed", {
-  title: __("allowed third party embed", "wb_block"),
+  title: __("allowed third party embed", "wb_block", "wb_block"),
   description: __("Add code from an allowed third party provider", "wb_block"),
   category: "wb-blocks",
   icon: "embed-generic",
   keywords: [
-    __("html", "third party embed", "smart survey"),
-    __("click", "wb_block"),
+    __("html", "wb_block"),
+    __("third party embed", "wb_block"),
+    __("smart survey", "wb_block"),
   ],
   attributes: {
     embedCode: {
@@ -30,74 +23,6 @@ registerBlockType("wb-blocks/allowed-third-party-embed", {
       default: "",
     },
   },
-  edit: (props) => {
-    const {
-      setAttributes,
-      attributes: { embedCode },
-      className,
-    } = props;
-
-    const [isPreview, setIsPreview] = useState(false);
-
-    const onChangeEmbedCode = (newEmbedCode) => {
-      setAttributes({
-        embedCode: newEmbedCode,
-      });
-
-      //reset to edit mode when code chnages so we don't display
-      //an outdated prewiew
-      //setIsPreview(false);
-    };
-
-    return [
-      <BlockControls>
-        <ToolbarGroup>
-          <ToolbarButton
-            isPressed={!isPreview}
-            onClick={() => {
-              setIsPreview(false);
-            }}
-          >
-            {__("Edit", "wb_block")}
-          </ToolbarButton>
-
-          <ToolbarButton
-            isPressed={isPreview}
-            disabled={!embedCode.trim()}
-            onClick={() => {
-              setIsPreview(true);
-            }}
-          >
-            {__("Preview", "wb_block")}
-          </ToolbarButton>
-        </ToolbarGroup>
-      </BlockControls>,
-
-      <div className={`wb-allowed-third-party-embed ${className || ""}`}>
-        {isPreview ? (
-          <div className="wb-allowed-third-party-embed__preview">
-            <SandBox
-              html={embedCode}
-              title={__("Third-party embed preview", "wb_block")}
-              type="embed"
-            />
-          </div>
-        ) : (
-          <div className="wb-allowed-third-party-embed__editor">
-            <TextareaControl
-              label={__("Third-party embed code", "wb_block")}
-              help={__(
-                "For now, paste some simple HTML to test the block.",
-                "wb_block",
-              )}
-              value={embedCode}
-              onChange={onChangeEmbedCode}
-              rows={10}
-            />
-          </div>
-        )}
-      </div>,
-    ];
-  },
+  edit: Edit,
   save: () => null,
 });
