+++ src/custom-blocks/allowed-third-party-embed/index.php
@@ -10,7 +10,7 @@
 /**
  * Phase 1: render simple html without validation
  * Phase 2: Validate provider
- * Phase 3: Validate provider
+ * Phase 3: Validate provider and html
  * Phase 4: Return only provider user and validated code
  */
 function wb_blocks_render_callback_allowed_third_party_embed( $attributes ) {
